requirements.txt
streamlit==1.35.0
langchain==0.1.20
requests==2.31.0
python-dotenv==1.0.1
dashscope>=1.10.0  # 通义千问官方SDK